requirements.txt
Django==4.0.6
djangorestframework-simplejwt==5.3.1
djangorestframework==3.13.1
gunicorn>=20.0.4
psycopg2-binary
